test/fix: create basic test cases to test the logic and GUI, fix small issues.
move the common utility logic to a separate folder.
create a folder for tests with the assembly definition file.
create tests for the functional core of the utility.
create tests for the utility's GUI logic.
add additional methods and properties to the main utility logic.
add new filters for selecting files via SFB (filters for 3d models, json/xml files)
fix caching of error and result messages in the utility logic - add resetting of messages
fix the position for the option to overwrite files, move it to the export settings section